Fix Kernel Panic related to loading events_df
In the first cell of the "Load review data and build interaction dataset" section, reading the JSON into a DataFrame reliably causes Kernel Panic (reproduced 3/3 times on an m5d.xlarge instance), believed to be due to overwhelming the memory available. Proposed solution is to load the dataset into a list of chunked smaller DataFrames then concatenate them once the dataset is fully loaded.