Remove non-required cell
The first cell of the "Get the Personalize API model Json and Personalize Boto3 Client" section references the downloading of a private Amazon internal file. This cell cannot be run by the general public as written and effectively does nothing anyways, as Personalize is a standard service accessible from the CLI and Boto3.